DSP-1349 Fix SSL support
Add setters for SSLSocketFactory and HostnameVerifier to Authenticator interface.
In both the KerberosAuthenticator & PseudoAuthenticator implementations, these
two members are initially set to the system defaults, but the AuthenticatedURL
instance which uses them can use the setters to ensure custom instances are
used. This is necessary when custom keystores, or hostname policies are required.
AuthenticatedURL may provided with an SSLSocketFactory & HostnameVerifier in its
constructor, which it then passes on to its Authenticator(s). Any connections that
the Authenticator(s) then open will use these helpers.