DSP-9181 Allow login on existing subject
The introduction of a SecurityManager in Cassandra 3.0 means that
Tomcat DispatchFilters run within a privileged context. As a result
there is already a Subject on the AccessContext that has not been
logged in. This patch also checks if there are any credentials on
the subject before deciding whether to login or not

Versions